app/BACKEND/weddingpictures/requirements.txt
boto3==1.7.49
botocore==1.10.49
dj-database-url==0.5.0
Django==2.0.7
django-cors-headers==2.3.0
django-heroku==0.3.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
docutils==0.14
gunicorn==19.9.0
jmespath==0.9.3
psycopg2==2.7.5
PyJWT==1.6.4
python-dateutil==2.7.3
pytz==2018.5
s3transfer==0.1.13
six==1.11.0
whitenoise==3.3.1
